# lib/api-zod/src/generated/types/createEcosystemInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Self-Healing Ecosystem Model API
 * OpenAPI spec version: 0.1.0
 */
import type { EcosystemType } from "./ecosystemType";
import type { Species } from "./species";

export interface CreateEcosystemInput {
  name: string;
  type: EcosystemType;
  /**
   * @minimum -10
   * @maximum 50
   */
  temperature: number;
  /**
   * @minimum 0
   * @maximum 5000
   */
  rainfall: number;
  /**
   * @minimum 0
   * @maximum 1
   */
  pollution: number;
  /**
   * @minimum 0
   * @maximum 1
   */
  deforestationRate: number;
  species: Species[];
}