# lib/api-zod/src/generated/types/ecosystem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Self-Healing Ecosystem Model API
 * OpenAPI spec version: 0.1.0
 */
import type { EcosystemInterventionsItem } from "./ecosystemInterventionsItem";
import type { EcosystemType } from "./ecosystemType";
import type { MLPredictions } from "./mLPredictions";
import type { Species } from "./species";

export interface Ecosystem {
  id: string;
  name: string;
  type: EcosystemType;
  temperature: number;
  rainfall: number;
  pollution: number;
  deforestationRate: number;
  species: Species[];
  predictions: MLPredictions;
  interventions: EcosystemInterventionsItem[];
  createdAt: string;
  updatedAt: string;
}