# lib/api-zod/src/generated/types/ecosystemType.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Self-Healing Ecosystem Model API
 * OpenAPI spec version: 0.1.0
 */

export type EcosystemType = (typeof EcosystemType)[keyof typeof EcosystemType];

export const EcosystemType = {
  forest: "forest",
  river: "river",
  grassland: "grassland",
} as const;