# lib/api-zod/src/generated/types/healthStatusEnum.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Self-Healing Ecosystem Model API
 * OpenAPI spec version: 0.1.0
 */

export type HealthStatusEnum =
  (typeof HealthStatusEnum)[keyof typeof HealthStatusEnum];

export const HealthStatusEnum = {
  Healthy: "Healthy",
  Stressed: "Stressed",
  Collapsing: "Collapsing",
} as const;