# lib/api-zod/src/generated/types/interventionInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Self-Healing Ecosystem Model API
 * OpenAPI spec version: 0.1.0
 */
import type { InterventionInputType } from "./interventionInputType";

export interface InterventionInput {
  type: InterventionInputType;
  /**
   * @minimum 0
   * @maximum 1
   */
  intensity: number;
}