# lib/api-zod/src/generated/types/mLPredictions.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Self-Healing Ecosystem Model API
 * OpenAPI spec version: 0.1.0
 */
import type { HealthStatusEnum } from "./healthStatusEnum";

export interface MLPredictions {
  healthStatus: HealthStatusEnum;
  /**
   * @minimum 0
   * @maximum 100
   */
  healthScore: number;
  recoveryTimeWeeks: number;
  /**
   * @minimum 0
   * @maximum 1
   */
  collapseRisk: number;
  /**
   * @minimum 0
   * @maximum 1
   */
  stabilityIndex: number;
  recommendations: string[];
}