# lib/api-zod/src/generated/types/simulationDataPoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Self-Healing Ecosystem Model API
 * OpenAPI spec version: 0.1.0
 */

export interface SimulationDataPoint {
  week: number;
  plants: number;
  herbivores: number;
  predators: number;
  healthScore: number;
}