# lib/api-zod/src/generated/types/simulationResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Self-Healing Ecosystem Model API
 * OpenAPI spec version: 0.1.0
 */
import type { MLPredictions } from "./mLPredictions";
import type { SimulationDataPoint } from "./simulationDataPoint";

export interface SimulationResult {
  id: string;
  ecosystemId: string;
  weeks: number;
  dataPoints: SimulationDataPoint[];
  finalPredictions: MLPredictions;
  runAt: string;
}