# lib/api-zod/src/generated/types/simulationRunInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Self-Healing Ecosystem Model API
 * OpenAPI spec version: 0.1.0
 */

export interface SimulationRunInput {
  /**
   * @minimum 1
   * @maximum 520
   */
  weeks?: number;
}