# lib/api-zod/src/generated/types/species.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Self-Healing Ecosystem Model API
 * OpenAPI spec version: 0.1.0
 */
import type { SpeciesType } from "./speciesType";

export interface Species {
  id: string;
  name: string;
  type: SpeciesType;
  initialPopulation: number;
}